big cleanup

+++ networkConstants.h
@@ -3,10 +3,12 @@
 #ifndef networkConstants_h 
 #define networkConstants_h networkConstants_h
 
+#include <stdint.h>
+
 namespace conedy 
 {  
   
-typedef int networkElementType;
+typedef uint16_t networkElementType;
   
   
   
@@ -36,9 +38,7 @@ const networkElementType _gaussianHR_=36;
 const networkElementType _gaussianNapKKm_=37;
 const networkElementType _gaussianNapK_=38;
 const networkElementType _gslOdeNode_=39;
-//const networkElementType _hindmarshRose_=40;
 const networkElementType _hopfNode_=42;
-//const networkElementType _integrateAndFire_=43;
 const networkElementType _izhikevichMap_=44;
 const networkElementType _kuramotoEdge_=45;
 const networkElementType _membranePotentialOfIFQ_=47;
@@ -51,7 +51,6 @@ const networkElementType _node_=53;
 const networkElementType _odeNode_=54;
 const networkElementType _pcoBase_=56;
 const networkElementType _staticComponent_=57;
-//const networkElementType _pcoDelay_=57;
 const networkElementType _pcoExponential_=58;
 const networkElementType _pcoRealIFNeuronDelay_=63;
 const networkElementType _pcoTrapez_=64;
@@ -90,7 +89,6 @@ const networkElementType _createNetwork_=97;
 const networkElementType _stdOdeIntegrator_=98;
 const networkElementType _stdSdeIntegrator_=99;
 
-//#include "generatedNetworkConstants.h" ;
 }
